🛠️ refactor: API 명세에 맞춰 Long 타입을 Integer로 변경

+++ src/main/java/prography/spring/pingpong/domain/game/controller/GameController.java
@@ -25,7 +25,7 @@ public class GameController {
     @Operation(summary = "게임 시작 API", description = "호스트가 게임을 시작합니다.")
     @PutMapping("/start/{roomId}")
     public ApiResponse<Void> startGame(
-            @PathVariable Long roomId,
+            @PathVariable Integer roomId,
             @RequestBody GameStartRequestDto requestDto
             ) {
         return gameService.startGame(roomId, requestDto);
