+++ src/main/java/prography/spring/pingpong/domain/game/event/GameStartedEvent.java
@@ -1,4 +1,4 @@
 package prography.spring.pingpong.domain.game.event;
 
-public record GameStartedEvent(Long roomId) {
+public record GameStartedEvent(Integer roomId) {
 }
