+++ src/main/java/prography/spring/pingpong/domain/userroom/model/dto/RoomJoinRequestDto.java
@@ -1,4 +1,4 @@
 package prography.spring.pingpong.domain.userroom.model.dto;
 
-public record RoomJoinRequestDto(Long userId) {
+public record RoomJoinRequestDto(int userId) {
 }
